// src/app/core/api/model/publicRepoSearchResultDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PublicRepoSearchResultDto { 
    /**
     * Repository name
     */
    name: string;
    /**
     * Full repository name (owner/repo)
     */
    full_name: string;
    /**
     * Repository description
     */
    description: string | null;
    /**
     * Number of stars
     */
    stars: number;
    /**
     * Primary language
     */
    language: string | null;
    /**
     * Default branch name
     */
    default_branch: string;
    /**
     * HTTPS clone URL
     */
    clone_url: string;
    /**
     * Repository HTML URL
     */
    html_url: string;
    /**
     * Always false for public repos
     */
    is_private: boolean;
}